Update test file organization, update pom.xml to allow tests to run
