+++ gettyimagesapi-sdk/src/test/java/DownloadImagesTest.java
@@ -1,7 +1,9 @@
-package unittests.Downloads;
+
 
 import com.gettyimages.ApiClient;
-import com.gettyimages.Downloads.DownloadVideos;
+import com.gettyimages.Downloads.DownloadImages;
+import com.gettyimages.Filters.FileType;
+import com.gettyimages.Filters.ProductType;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
@@ -17,7 +19,7 @@ import static org.mockserver.integration.ClientAndServer.startClientAndServer;
 import static org.mockserver.model.HttpRequest.request;
 import static org.mockserver.model.HttpResponse.response;
 
-public class DownloadVideosTests {
+public class DownloadImagesTest {
     private static ClientAndServer mockServer;
 
     @BeforeAll
@@ -44,7 +46,27 @@ public class DownloadVideosTests {
         client.when(
                 request()
                         .withMethod("POST")
-                        .withPath("/downloads/videos/12345")
+                        .withPath("/downloads/images/12345")
+                        .withQueryStringParameters(
+                                new Parameter("file_type", "jpg"),
+                                new Parameter("auto_download", "false")
+                        )
+        )
+                .respond(response().withStatusCode(200).withBody("success"));
+        client.when(
+                request()
+                        .withMethod("POST")
+                        .withPath("/downloads/images/12345")
+                        .withQueryStringParameters(
+                                new Parameter("auto_download", "false"),
+                                new Parameter("height", "592")
+                        )
+        )
+                .respond(response().withStatusCode(200).withBody("success"));
+        client.when(
+                request()
+                        .withMethod("POST")
+                        .withPath("/downloads/images/12345")
                         .withQueryStringParameters(
                                 new Parameter("product_id", "9876"),
                                 new Parameter("auto_download", "false")
@@ -54,9 +76,9 @@ public class DownloadVideosTests {
         client.when(
                 request()
                         .withMethod("POST")
-                        .withPath("/downloads/videos/12345")
+                        .withPath("/downloads/images/12345")
                         .withQueryStringParameters(
-                                new Parameter("size", "size"),
+                                new Parameter("product_type", "easyaccess"),
                                 new Parameter("auto_download", "false")
                         )
         )
@@ -65,21 +87,41 @@ public class DownloadVideosTests {
     }
 
     @Test
-    void downloadVideosWithProductId() throws Exception {
+    void downloadImagesWithFileType() throws Exception {
         ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
-        DownloadVideos downloadVideos = client.downloadvideos()
-                .withId("12345").withProductId(9876);
-        String result = downloadVideos.executeAsync();
+        DownloadImages downloadimages = client.downloadimages()
+                .withId("12345").withFileType(FileType.JPG);
+        String result = downloadimages.executeAsync();
         System.out.print(result);
         assertEquals("success", result);
     }
 
     @Test
-    void downloadVideosWithSize() throws Exception {
+    void downloadImagesWithHeight() throws Exception {
         ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
-        DownloadVideos downloadVideos = client.downloadvideos()
-                .withId("12345").withSize("size");
-        String result = downloadVideos.executeAsync();
+        DownloadImages downloadimages = client.downloadimages()
+                .withId("12345").withHeight("592");
+        String result = downloadimages.executeAsync();
+        System.out.print(result);
+        assertEquals("success", result);
+    }
+
+    @Test
+    void downloadImagesWithProductId() throws Exception {
+        ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
+        DownloadImages downloadimages = client.downloadimages()
+                .withId("12345").withProductId(9876);
+        String result = downloadimages.executeAsync();
+        System.out.print(result);
+        assertEquals("success", result);
+    }
+
+    @Test
+    void downloadImagesWithProductType() throws Exception {
+        ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
+        DownloadImages downloadimages = client.downloadimages()
+                .withId("12345").withProductType(ProductType.EASYACCESS);
+        String result = downloadimages.executeAsync();
         System.out.print(result);
         assertEquals("success", result);
     }
