+++ gettyimagesapi-sdk/src/test/java/SearchImagesEditorialTest.java
@@ -1,8 +1,7 @@
-package unittests.Search;
 
 import com.gettyimages.ApiClient;
 import com.gettyimages.Filters.*;
-import com.gettyimages.Search.SearchImages;
+import com.gettyimages.Search.SearchImagesEditorial;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
@@ -10,7 +9,6 @@ import org.junit.jupiter.api.Test;
 import org.mockserver.client.server.MockServerClient;
 import org.mockserver.integration.ClientAndServer;
 import org.mockserver.model.Parameter;
-import org.mozilla.javascript.EcmaError;
 
 import java.lang.reflect.Field;
 import java.util.Arrays;
@@ -21,7 +19,7 @@ import static org.mockserver.integration.ClientAndServer.startClientAndServer;
 import static org.mockserver.model.HttpRequest.request;
 import static org.mockserver.model.HttpResponse.response;
 
-public class SearchImagesTests {
+public class SearchImagesEditorialTest {
     private static ClientAndServer mockServer;
 
     @BeforeAll
@@ -48,7 +46,7 @@ public class SearchImagesTests {
         client.when(
                 request()
                         .withMethod("GET")
-                        .withPath("/search/images")
+                        .withPath("/search/images/editorial")
                         .withQueryStringParameters(
                                 new Parameter("age_of_people", "baby,child,adult")
                         )
@@ -57,7 +55,7 @@ public class SearchImagesTests {
         client.when(
                 request()
                         .withMethod("GET")
-                        .withPath("/search/images")
+                        .withPath("/search/images/editorial")
                         .withQueryStringParameters(
                                 new Parameter("artists", "roman makhmutov")
                         )
@@ -66,7 +64,7 @@ public class SearchImagesTests {
         client.when(
                 request()
                         .withMethod("GET")
-                        .withPath("/search/images")
+                        .withPath("/search/images/editorial")
                         .withQueryStringParameters(
                                 new Parameter("collection_codes", "WRI,ARF")
                         )
@@ -75,7 +73,7 @@ public class SearchImagesTests {
         client.when(
                 request()
                         .withMethod("GET")
-                        .withPath("/search/images")
+                        .withPath("/search/images/editorial")
                         .withQueryStringParameters(
                                 new Parameter("collections_filter_type", "exclude")
                         )
@@ -84,16 +82,7 @@ public class SearchImagesTests {
         client.when(
                 request()
                         .withMethod("GET")
-                        .withPath("/search/images")
-                        .withQueryStringParameters(
-                                new Parameter("color", "#002244")
-                        )
-        )
-                .respond(response().withStatusCode(200).withBody("success"));
-        client.when(
-                request()
-                        .withMethod("GET")
-                        .withPath("/search/images")
+                        .withPath("/search/images/editorial")
                         .withQueryStringParameters(
                                 new Parameter("compositions", "abstract,headshot")
                         )
@@ -102,7 +91,16 @@ public class SearchImagesTests {
         client.when(
                 request()
                         .withMethod("GET")
-                        .withPath("/search/images")
+                        .withPath("/search/images/editorial")
+                        .withQueryStringParameters(
+                                new Parameter("editorial_segments", "archival,publicity")
+                        )
+        )
+                .respond(response().withStatusCode(200).withBody("success"));
+        client.when(
+                request()
+                        .withMethod("GET")
+                        .withPath("/search/images/editorial")
                         .withQueryStringParameters(
                                 new Parameter("embed_content_only", "true")
                         )
@@ -111,7 +109,25 @@ public class SearchImagesTests {
         client.when(
                 request()
                         .withMethod("GET")
-                        .withPath("/search/images")
+                        .withPath("/search/images/editorial")
+                        .withQueryStringParameters(
+                                new Parameter("end_date", "2015-04-01")
+                        )
+        )
+                .respond(response().withStatusCode(200).withBody("success"));
+        client.when(
+                request()
+                        .withMethod("GET")
+                        .withPath("/search/images/editorial")
+                        .withQueryStringParameters(
+                                new Parameter("entity_uris", "uri1,uri2")
+                        )
+        )
+                .respond(response().withStatusCode(200).withBody("success"));
+        client.when(
+                request()
+                        .withMethod("GET")
+                        .withPath("/search/images/editorial")
                         .withQueryStringParameters(
                                 new Parameter("ethnicity", "black,japanese")
                         )
@@ -120,7 +136,7 @@ public class SearchImagesTests {
         client.when(
                 request()
                         .withMethod("GET")
-                        .withPath("/search/images")
+                        .withPath("/search/images/editorial")
                         .withQueryStringParameters(
                                 new Parameter("event_ids", "123,456,789")
                         )
@@ -129,7 +145,7 @@ public class SearchImagesTests {
         client.when(
                 request()
                         .withMethod("GET")
-                        .withPath("/search/images")
+                        .withPath("/search/images/editorial")
                         .withQueryStringParameters(
                                 new Parameter("exclude_nudity", "true")
                         )
@@ -138,7 +154,7 @@ public class SearchImagesTests {
         client.when(
                 request()
                         .withMethod("GET")
-                        .withPath("/search/images")
+                        .withPath("/search/images/editorial")
                         .withQueryStringParameters(
                                 new Parameter("fields", "asset_family,id,uri_oembed")
                         )
@@ -147,7 +163,7 @@ public class SearchImagesTests {
         client.when(
                 request()
                         .withMethod("GET")
-                        .withPath("/search/images")
+                        .withPath("/search/images/editorial")
                         .withQueryStringParameters(
                                 new Parameter("file_types", "eps,jpg")
                         )
@@ -156,7 +172,7 @@ public class SearchImagesTests {
         client.when(
                 request()
                         .withMethod("GET")
-                        .withPath("/search/images")
+                        .withPath("/search/images/editorial")
                         .withQueryStringParameters(
                                 new Parameter("graphical_styles", "fine_art,illustration")
                         )
@@ -165,7 +181,7 @@ public class SearchImagesTests {
         client.when(
                 request()
                         .withMethod("GET")
-                        .withPath("/search/images")
+                        .withPath("/search/images/editorial")
                         .withQueryStringParameters(
                                 new Parameter("keyword_ids", "1111,2222,3333")
                         )
@@ -174,16 +190,16 @@ public class SearchImagesTests {
         client.when(
                 request()
                         .withMethod("GET")
-                        .withPath("/search/images")
+                        .withPath("/search/images/editorial")
                         .withQueryStringParameters(
-                                new Parameter("license_models", "rightsmanaged")
+                                new Parameter("minimum_quality_rank", "1")
                         )
         )
                 .respond(response().withStatusCode(200).withBody("success"));
         client.when(
                 request()
                         .withMethod("GET")
-                        .withPath("/search/images")
+                        .withPath("/search/images/editorial")
                         .withQueryStringParameters(
                                 new Parameter("minimum_size", "small")
                         )
@@ -192,7 +208,7 @@ public class SearchImagesTests {
         client.when(
                 request()
                         .withMethod("GET")
-                        .withPath("/search/images")
+                        .withPath("/search/images/editorial")
                         .withQueryStringParameters(
                                 new Parameter("number_of_people", "one,group")
                         )
@@ -201,7 +217,7 @@ public class SearchImagesTests {
         client.when(
                 request()
                         .withMethod("GET")
-                        .withPath("/search/images")
+                        .withPath("/search/images/editorial")
                         .withQueryStringParameters(
                                 new Parameter("orientations", "Horizontal,Square")
                         )
@@ -210,7 +226,7 @@ public class SearchImagesTests {
         client.when(
                 request()
                         .withMethod("GET")
-                        .withPath("/search/images")
+                        .withPath("/search/images/editorial")
                         .withQueryStringParameters(
                                 new Parameter("page", "3")
                         )
@@ -219,7 +235,7 @@ public class SearchImagesTests {
         client.when(
                 request()
                         .withMethod("GET")
-                        .withPath("/search/images")
+                        .withPath("/search/images/editorial")
                         .withQueryStringParameters(
                                 new Parameter("page_size", "50")
                         )
@@ -228,7 +244,7 @@ public class SearchImagesTests {
         client.when(
                 request()
                         .withMethod("GET")
-                        .withPath("/search/images")
+                        .withPath("/search/images/editorial")
                         .withQueryStringParameters(
                                 new Parameter("phrase", "cat")
                         )
@@ -237,16 +253,7 @@ public class SearchImagesTests {
         client.when(
                 request()
                         .withMethod("GET")
-                        .withPath("/search/images")
-                        .withQueryStringParameters(
-                                new Parameter("prestige_content_only", "true")
-                        )
-        )
-                .respond(response().withStatusCode(200).withBody("success"));
-        client.when(
-                request()
-                        .withMethod("GET")
-                        .withPath("/search/images")
+                        .withPath("/search/images/editorial")
                         .withQueryStringParameters(
                                 new Parameter("product_types", "easyaccess,editorialsubscription")
                         )
@@ -255,7 +262,7 @@ public class SearchImagesTests {
         client.when(
                 request()
                         .withMethod("GET")
-                        .withPath("/search/images")
+                        .withPath("/search/images/editorial")
                         .withQueryStringParameters(
                                 new Parameter("sort_order", "newest")
                         )
@@ -264,18 +271,27 @@ public class SearchImagesTests {
         client.when(
                 request()
                         .withMethod("GET")
-                        .withPath("/search/images")
+                        .withPath("/search/images/editorial")
                         .withQueryStringParameters(
                                 new Parameter("specific_people", "Reggie Jackson")
+                        )
+        )
+                .respond(response().withStatusCode(200).withBody("success"));
+        client.when(
+                request()
+                        .withMethod("GET")
+                        .withPath("/search/images/editorial")
+                        .withQueryStringParameters(
+                                new Parameter("start_date", "2015-04-01")
                         )
         )
                 .respond(response().withStatusCode(200).withBody("success"));
     }
 
     @Test
-    void searchImagesWithAgeOfPeople() throws Exception {
+    void searchImagesEditorialWithAgeOfPeople() throws Exception {
         ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
-        SearchImages search = client.searchimages()
+        SearchImagesEditorial search = client.searchimageseditorial()
                 .withAgeOfPeople(EnumSet.of(AgeOfPeople.ADULT)).withAgeOfPeople(EnumSet.of(AgeOfPeople.CHILD, AgeOfPeople.BABY));
         String result = search.executeAsync();
         System.out.print(result);
@@ -283,9 +299,9 @@ public class SearchImagesTests {
     }
 
     @Test
-    void searchImagesWithArtist() throws Exception {
+    void searchImagesEditorialWithArtist() throws Exception {
         ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
-        SearchImages search = client.searchimages()
+        SearchImagesEditorial search = client.searchimageseditorial()
                 .withArtists(Arrays.asList("roman makhmutov"));
         String result = search.executeAsync();
         System.out.print(result);
@@ -293,9 +309,9 @@ public class SearchImagesTests {
     }
 
     @Test
-    void searchImagesWithCollectionCodes() throws Exception {
+    void searchImagesEditorialWithCollectionCodes() throws Exception {
         ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
-        SearchImages search = client.searchimages()
+        SearchImagesEditorial search = client.searchimageseditorial()
                 .withCollectionCodes(Arrays.asList("WRI", "ARF"));
         String result = search.executeAsync();
         System.out.print(result);
@@ -303,30 +319,20 @@ public class SearchImagesTests {
     }
 
     @Test
-    void searchImagesWithCollectionFilter() throws Exception {
+    void searchImagesEditorialWithCollectionFilter() throws Exception {
         ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
-        SearchImages search = client.searchimages()
+        SearchImagesEditorial search = client.searchimageseditorial()
                 .withCollectionFilterType(CollectionFilter.EXCLUDE);
         String result = search.executeAsync();
         System.out.print(result);
         assertEquals("success", result);
     }
 
-    @Test
-    void searchImagesWithColor() throws Exception {
-        ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
-        SearchImages search = client.searchimages()
-                .withColor("#002244");
-        String result = search.executeAsync();
-        System.out.print(result);
-        assertEquals("success", result);
-    }
-
 
     @Test
-    void searchImagesWithCompositions() throws Exception {
+    void searchImagesEditorialWithComposition() throws Exception {
         ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
-        SearchImages search = client.searchimages()
+        SearchImagesEditorial search = client.searchimageseditorial()
                 .withCompositions(EnumSet.of(Compositions.HEADSHOT, Compositions.ABSTRACT));
         String result = search.executeAsync();
         System.out.print(result);
@@ -334,9 +340,19 @@ public class SearchImagesTests {
     }
 
     @Test
-    void searchImagesWithEmbedContentOnly() throws Exception {
+    void searchImagesEditorialWithEditorialSegments() throws Exception {
         ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
-        SearchImages search = client.searchimages()
+        SearchImagesEditorial search = client.searchimageseditorial()
+                .withEditorialSegments(EnumSet.of(EditorialSegment.ARCHIVAL, EditorialSegment.PUBLICITY));
+        String result = search.executeAsync();
+        System.out.print(result);
+        assertEquals("success", result);
+    }
+
+    @Test
+    void searchImagesEditorialWithEmbedContentOnly() throws Exception {
+        ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
+        SearchImagesEditorial search = client.searchimageseditorial()
                 .withEmbedContentOnly(true);
         String result = search.executeAsync();
         System.out.print(result);
@@ -344,9 +360,29 @@ public class SearchImagesTests {
     }
 
     @Test
-    void searchImagesWithEthnicity() throws Exception {
+    void searchImagesEditorialWithEndDate() throws Exception {
         ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
-        SearchImages search = client.searchimages()
+        SearchImagesEditorial search = client.searchimageseditorial()
+                .withEndDate("2015-04-01");
+        String result = search.executeAsync();
+        System.out.print(result);
+        assertEquals("success", result);
+    }
+
+    @Test
+    void searchImagesEditorialWithEntityUris() throws Exception{
+        ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
+        SearchImagesEditorial search = client.searchimageseditorial()
+                .withEntityUris(Arrays.asList("uri1", "uri2"));
+        String result = search.executeAsync();
+        System.out.print(result);
+        assertEquals("success", result);
+    }
+
+    @Test
+    void searchImagesEditorialWithEthnicity() throws Exception {
+        ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
+        SearchImagesEditorial search = client.searchimageseditorial()
                 .withEthnicity(EnumSet.of(Ethnicity.BLACK, Ethnicity.JAPANESE));
         String result = search.executeAsync();
         System.out.print(result);
@@ -354,9 +390,9 @@ public class SearchImagesTests {
     }
 
     @Test
-    void searchImagesWithEventIds() throws Exception {
+    void searchImagesEditorialWithEventIds() throws Exception {
         ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
-        SearchImages search = client.searchimages()
+        SearchImagesEditorial search = client.searchimageseditorial()
                 .withEventIds(Arrays.asList(123, 456, 789));
         String result = search.executeAsync();
         System.out.print(result);
@@ -364,9 +400,9 @@ public class SearchImagesTests {
     }
 
     @Test
-    void searchImagesWithExcludeNudity() throws Exception {
+    void searchImagesEditorialWithExcludeNudity() throws Exception {
         ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
-        SearchImages search = client.searchimages()
+        SearchImagesEditorial search = client.searchimageseditorial()
                 .withExcludeNudity(true);
         String result = search.executeAsync();
         System.out.print(result);
@@ -374,9 +410,9 @@ public class SearchImagesTests {
     }
 
     @Test
-    void searchImagesWithResponseFields() throws Exception {
+    void searchImagesEditorialWithResponseFields() throws Exception {
         ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
-        SearchImages search = client.searchimages()
+        SearchImagesEditorial search = client.searchimageseditorial()
                 .withResponseFields(Arrays.asList("asset_family", "id", "uri_oembed"));
         String result = search.executeAsync();
         System.out.print(result);
@@ -384,9 +420,9 @@ public class SearchImagesTests {
     }
 
     @Test
-    void searchImagesWithFileTypes() throws Exception {
+    void searchImagesEditorialWithFileTypes() throws Exception {
         ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
-        SearchImages search = client.searchimages()
+        SearchImagesEditorial search = client.searchimageseditorial()
                 .withFileTypes(EnumSet.of(FileType.EPS, FileType.JPG));
         String result = search.executeAsync();
         System.out.print(result);
@@ -394,9 +430,9 @@ public class SearchImagesTests {
     }
 
    @Test
-    void searchImagesWithGraphicalStyles() throws Exception {
+    void searchImagesEditorialWithGraphicalStyles() throws Exception {
         ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
-        SearchImages search = client.searchimages()
+        SearchImagesEditorial search = client.searchimageseditorial()
                 .withGraphicalStyles(EnumSet.of(GraphicalStyles.FINE_ART, GraphicalStyles.ILLUSTRATION));
         String result = search.executeAsync();
         System.out.print(result);
@@ -404,9 +440,9 @@ public class SearchImagesTests {
     }
 
     @Test
-    void searchImagesWithKeywordIds() throws Exception {
+    void searchImagesEditorialWithKeywordIds() throws Exception {
         ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
-        SearchImages search = client.searchimages()
+        SearchImagesEditorial search = client.searchimageseditorial()
                 .withKeywordIds(Arrays.asList(1111,2222,3333));
         String result = search.executeAsync();
         System.out.print(result);
@@ -414,29 +450,29 @@ public class SearchImagesTests {
     }
 
     @Test
-    void searchImagesWithLicenseModels() throws Exception {
+    void searchImagesEditorialWithMinimumQualityRank() throws Exception {
         ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
-        SearchImages search = client.searchimages()
-                .withLicenseModels(EnumSet.of(LicenseModel.RIGHTS_MANAGED));
+        SearchImagesEditorial search = client.searchimageseditorial()
+                .withMinimumQualityRank(MinimumQualityRank.ONE);
         String result = search.executeAsync();
         System.out.print(result);
         assertEquals("success", result);
     }
 
     @Test
-    void searchImagesWithMinimumSize() throws Exception {
+    void searchImagesEditorialWithMinimumSize() throws Exception {
         ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
-        SearchImages search = client.searchimages()
-            .withMinimumSize(MinimumSize.SMALL);
+        SearchImagesEditorial search = client.searchimageseditorial()
+                .withMinimumSize(MinimumSize.SMALL);
         String result = search.executeAsync();
         System.out.print(result);
         assertEquals("success", result);
     }
 
     @Test
-    void searchImagesWithNumberOfPeople() throws Exception {
+    void searchImagesEditorialWithNumberOfPeople() throws Exception {
         ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
-        SearchImages search = client.searchimages()
+        SearchImagesEditorial search = client.searchimageseditorial()
                 .withNumberOfPeople(EnumSet.of(NumberOfPeople.GROUP, NumberOfPeople.ONE));
         String result = search.executeAsync();
         System.out.print(result);
@@ -444,9 +480,9 @@ public class SearchImagesTests {
     }
 
     @Test
-    void searchImagesWithOrientations() throws Exception {
+    void searchImagesEditorialWithOrientations() throws Exception {
         ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
-        SearchImages search = client.searchimages()
+        SearchImagesEditorial search = client.searchimageseditorial()
                 .withOrientations(EnumSet.of(Orientation.HORIZONTAL, Orientation.SQUARE));
         String result = search.executeAsync();
         System.out.print(result);
@@ -454,9 +490,9 @@ public class SearchImagesTests {
     }
 
     @Test
-    void searchImagesWithPage() throws Exception {
+    void searchImagesEditorialWithPage() throws Exception {
         ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
-        SearchImages search = client.searchimages()
+        SearchImagesEditorial search = client.searchimageseditorial()
                 .withPage(3);
         String result = search.executeAsync();
         System.out.print(result);
@@ -464,9 +500,9 @@ public class SearchImagesTests {
     }
 
     @Test
-    void searchImagesWithPageSize() throws Exception {
+    void searchImagesEditorialWithPageSize() throws Exception {
         ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
-        SearchImages search = client.searchimages()
+        SearchImagesEditorial search = client.searchimageseditorial()
                 .withPageSize(50);
         String result = search.executeAsync();
         System.out.print(result);
@@ -474,9 +510,9 @@ public class SearchImagesTests {
     }
 
     @Test
-    void searchImagesWithPhrase() throws Exception {
+    void searchImagesEditorialWithPhrase() throws Exception {
         ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
-        SearchImages search = client.searchimages()
+        SearchImagesEditorial search = client.searchimageseditorial()
                 .withPhrase("cat");
         String result = search.executeAsync();
         System.out.print(result);
@@ -484,19 +520,9 @@ public class SearchImagesTests {
     }
 
     @Test
-    void searchImagesWithPrestigeContentOnly() throws Exception {
+    void searchImagesEditorialWithProductTypes() throws Exception {
         ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
-        SearchImages search = client.searchimages()
-                .withPrestigeContentOnly(true);
-        String result = search.executeAsync();
-        System.out.print(result);
-        assertEquals("success", result);
-    }
-
-    @Test
-    void searchImagesWithProductTypes() throws Exception {
-        ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
-        SearchImages search = client.searchimages()
+        SearchImagesEditorial search = client.searchimageseditorial()
                 .withProductTypes(EnumSet.of(ProductType.EASYACCESS, ProductType.EDITORIALSUBSCRIPTION));
         String result = search.executeAsync();
         System.out.print(result);
@@ -504,9 +530,9 @@ public class SearchImagesTests {
     }
 
     @Test
-    void searchImagesWithSortOrder() throws Exception {
+    void searchImagesEditorialWithSortOrder() throws Exception {
         ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
-        SearchImages search = client.searchimages()
+        SearchImagesEditorial search = client.searchimageseditorial()
                 .withSortOrder(SortOrder.NEWEST);
         String result = search.executeAsync();
         System.out.print(result);
@@ -514,10 +540,21 @@ public class SearchImagesTests {
     }
 
     @Test
-    void searchImagesWithSpecificPeople() throws Exception {
+    void searchImagesEditorialWithSpecificPeople() throws Exception {
         ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
-        SearchImages search = client.searchimages()
+        SearchImagesEditorial search = client.searchimageseditorial()
                 .withSpecificPeople(Arrays.asList("Reggie Jackson"));
+        String result = search.executeAsync();
+        System.out.print(result);
+        assertEquals("success", result);
+    }
+
+
+    @Test
+    void searchImagesEditorialWithStartDate() throws Exception {
+        ApiClient client = ApiClient.GetApiClientWithClientCredentials("apiKey", "apiSecret");
+        SearchImagesEditorial search = client.searchimageseditorial()
+                .withStartDate("2015-04-01");
         String result = search.executeAsync();
         System.out.print(result);
         assertEquals("success", result);
